pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azr"
version = "0.1.0"
description = "AZR RLHF training harness with GRPO/QLoRA and tool sandbox"
authors = [{ name = "AZR Research", email = "devnull@example.com" }]
requires-python = ">=3.9"
dependencies = [
  "transformers>=4.43.0",
  "trl>=0.23.0",
  "peft>=0.11.1",
  "accelerate>=0.33.0",
  "datasets>=2.19.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.1",
  "tqdm",
  "requests",
]

[project.scripts]
azr-train = "azr.cli:train_cmd"
azr-eval = "azr.cli:eval_cmd"
azr-gen = "azr.cli:gen_cmd"

